Refactor http response checksum mutation intercept
